FEAT: 반품 목록페이지, 반품 상세페이지, 반품신청시 주문상세페이지에 현황 반영

+++ backend/src/main/java/com/erp/backend/disposal/mapper/DisposalMapper.java
@@ -14,13 +14,13 @@ public interface DisposalMapper {
 
     Integer getCurrentDisposalDetailSeq();
 
-    Integer insertDisposalRequest(@Param("disposalId") int disposalId, @Param("reason") String reason, @Param("empId") int empId);
+    Integer insertDisposalRequest(@Param("disposalId") int disposalId, @Param("reason") String reason, @Param("empId") long empId);
 
     String findDisposalStatus(int disposalId);
 
     Integer insertDisposalDetail(DisposalDetailVO disposalDetailVO);
 
-    Integer approveDisposalRequest(@Param("approvedBy") int empId, @Param("disposalId") int disposalId);
+    Integer approveDisposalRequest(@Param("approvedBy") long empId, @Param("disposalId") int disposalId);
 
     Integer rejectDisposalRequest(@Param("rejectReason") String rejectReason, @Param("disposalId") int disposalId);
